requirements.txt
langchain_community
langchain_google_genai
langchain_pinecone
langchain_groq
pypdf
faiss-cpu
streamlit
